Move Lua to the LuaScript
